Updated PlaceRow tests.
