+++ assets/tests/components/placeRow.test.tsx
@@ -3,7 +3,6 @@ import { fireEvent, render } from "@testing-library/react";
 import PlaceRow from "../../js/components/Dashboard/PlaceRow";
 import { Accordion } from "react-bootstrap";
 import { Place } from "../../js/models/place";
-import { ScreenplayProvider } from "../../js/hooks/useScreenplayContext";
 
 beforeAll(() => {
   const app = document.createElement("div");
@@ -12,46 +11,6 @@ beforeAll(() => {
 });
 
 describe("PlaceRow", () => {
-  test("opens when clicked", async () => {
-    const place: Place = {
-      id: "place-stop1",
-      name: "Place Name1",
-      routes: ["CR", "Red", "Green-B"],
-      status: "Auto",
-      screens: [
-        { id: "1111", type: "dup", disabled: false },
-        { id: "2222", type: "solari", disabled: false },
-        { id: "3333", type: "bus_shelter_v2", disabled: false },
-      ],
-    };
-
-    const handleClick = jest.fn();
-
-    const { getByTestId, getByAltText, queryByAltText } = render(
-      <ScreenplayProvider>
-        <Accordion>
-          <PlaceRow
-            place={place}
-            eventKey="0"
-            onClick={handleClick}
-            defaultSort
-          />
-        </Accordion>
-      </ScreenplayProvider>
-    );
-
-    expect(getByTestId("place-row").className).toBe("place-row");
-    fireEvent.click(getByTestId("place-row-header"));
-    expect(handleClick).toHaveBeenCalled();
-    expect(getByTestId("place-row").className).toBe("place-row open");
-    expect(getByTestId("place-screen-types").textContent).toBe(
-      "DUP·Bus Shelter·Solari"
-    );
-    expect(getByTestId("place-status").textContent).toBe("Auto");
-    expect(getByAltText("Green-B")).toBeInTheDocument();
-    expect(queryByAltText("Green")).toBeNull();
-  });
-
   test("shows no screens", async () => {
     const place: Place = {
       id: "place-stop1",
@@ -70,6 +29,8 @@ describe("PlaceRow", () => {
           eventKey="0"
           onClick={handleClick}
           defaultSort
+          disabled
+          variant="accordion"
         />
       </Accordion>
     );
